Create Speaker.m and Speakers.mat
Decided that rather than having speaker parameters defined in get_speaker_spectrogram.m, they should be defined in a separate .mat file that can then be loaded. I decided to create a speaker class, and have all speakers saved in an object array that can be loaded by get_speaker_spectrogram.m. Also, the function for checking that a speaker's paramters are compatible with the requested stimulus is now method of the speaker class; the idea is that get_speaker_spectrogram will load the Speaker objects, find the one whose name matches that speaker name given as an input argument, then call that speaker's validation method with the requested stimulus parameters as inputs.